+++ NEWS.md
@@ -1,3 +1,8 @@
+# aion 1.4.0
+## New classes and methods
+* Add `as_graph()` to compute an interval graph.
+* Add `get_calendar()` and `set_calendar()` to get and set the default calendar.
+
 # aion 1.3.0
 ## New classes and methods
 * Add `calendar_year()` to get the average length of the year of a calendar (in solar days).